# docs/api/swagger.yaml
openapi: 3.0.0
info:
 title: Cheddar API
 version: 1.0.0
 description: API documentation for Cheddar social media scheduler
servers:
 - url: /api
   description: API base URL
components:
 securitySchemes:
  sessionAuth:
   type: apiKey
   in: cookie
   name: __Secure-next-auth.session-token
   description: Session token cookie set by NextAuth.js
 schemas:
  Task:
   type: object
   required:
    - type
    - priority
   properties:
    id:
     type: string
     example: 507f1f77bcf86cd799439011
    type:
     type: string
     enum: [copy, image, review]
     example: copy
    priority:
     type: integer
     minimum: 1
     maximum: 3
     example: 2
    status:
     type: string
     enum: [pending, processing, completed, failed]
     example: pending
    createdAt:
     type: string
     format: date-time
    updatedAt:
     type: string
     format: date-time
  Provider:
   type: object
   required:
    - name
    - capabilities
   properties:
    id:
     type: string
     example: 507f1f77bcf86cd799439011
    name:
     type: string
     example: Twitter API
    capabilities:
     type: array
     items:
      type: string
     example: [post, schedule, analytics]
    status:
     type: string
     enum: [active, degraded, inactive]
     example: active
    createdAt:
     type: string
     format: date-time
    updatedAt:
     type: string
     format: date-time
  Error:
   type: object
   properties:
    message:
     type: string
     example: Error message
    errors:
     type: array
     items:
      type: object
      properties:
       path:
        type: array
        items:
         type: string
       message:
        type: string
paths:
 /tasks:
  post:
   summary: Create new task
   tags: [Tasks]
   security:
    - sessionAuth: []
   requestBody:
    required: true
    content:
     application/json:
      schema:
       type: object
       properties:
        type:
         type: string
         enum: [copy, image, review]
        priority:
         type: integer
         minimum: 1
         maximum: 3
   responses:
    '201':
     description: Task created successfully
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Task'
    '400':
     description: Invalid request
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
    '401':
     description: Unauthorized
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
 /tasks/{id}:
  get:
   summary: Get task by ID
   tags: [Tasks]
   security:
    - sessionAuth: []
   parameters:
    - name: id
      in: path
      required: true
      schema:
       type: string
   responses:
    '200':
     description: Task details
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Task'
    '404':
     description: Task not found
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
    '401':
     description: Unauthorized
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'

  patch:
   summary: Update task
   tags: [Tasks]
   security:
    - sessionAuth: []
   parameters:
    - name: id
      in: path
      required: true
      schema:
       type: string
   requestBody:
    required: true
    content:
     application/json:
      schema:
       $ref: '#/components/schemas/Task'
   responses:
    '200':
     description: Task updated successfully
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Task'
    '404':
     description: Task not found
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
    '401':
     description: Unauthorized
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
 /providers:
  get:
   summary: List all providers
   tags: [Providers]
   security:
    - sessionAuth: []
   responses:
    '200':
     description: List of providers
     content:
      application/json:
       schema:
        type: array
        items:
         $ref: '#/components/schemas/Provider'
    '401':
     description: Unauthorized
     content:
      application/json:
       schema:
        $ref: '#/components/schemas/Error'
...